# app/data/demo_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Manager API",
    "description": "A simple task management API used as the built-in Demo Mode fixture for MCP Server Builder.",
    "version": "1.2.0"
  },
  "servers": [
    { "url": "https://api.demo-taskmanager.dev/v1" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "listProjects",
        "summary": "List projects",
        "description": "Returns all projects.",
        "tags": ["projects"],
        "parameters": [],
        "responses": {
          "200": { "description": "A list of projects", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } } }
        }
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create project",
        "description": "Creates a new project.",
        "tags": ["projects"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } }, "required": ["name"] } } }
        },
        "responses": {
          "201": { "description": "Project created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } }
        }
      }
    },
    "/projects/{id}": {
      "delete": {
        "operationId": "deleteProject",
        "summary": "Delete project",
        "description": "Permanently deletes a project and all of its tasks.",
        "tags": ["projects"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Project deleted" }
        }
      }
    }
  }
}